+++ languages/pl/tl_content.php
@@ -14,6 +14,7 @@ $GLOBALS['TL_LANG']['tl_content']['closingTags'] = array('Domykanie znaczników'
 $GLOBALS['TL_LANG']['tl_content']['wrapperTagsTag'] = array('Znacznik', 'Wybierz znacznik');
 $GLOBALS['TL_LANG']['tl_content']['wrapperTagsClass'] = array('Class', 'Wpisz klasy');
 $GLOBALS['TL_LANG']['tl_content']['wrapperTagsId'] = array('Id', 'Wpisz wartość id');
+$GLOBALS['TL_LANG']['tl_content']['wrapperTagsStyle'] = array('Style', 'Wpisz css-owe style');
 $GLOBALS['TL_LANG']['tl_content']['wrappertags_legend'] = 'Ustawienia HTML-owych znaczników';